save change with locaStoreId store

+++ my-shop/src/app/admin/products/page.tsx
@@ -6,6 +6,7 @@ import { Button, Box } from '@chakra-ui/react'
 //@ts-ignore
 import { AdminHeader } from '@/app/Components/AdminHeader'
 import { useAppContext } from '@/app/Core/Context'
+import { SpinnerComponent } from '@/app/Components/Spinner'
 import { AdminCategoriesComponent } from '@/app/Components/AdminCategories'
 import '../../main.scss'
 import Link from '../../../../node_modules/next/link'
@@ -14,31 +15,44 @@ import { Products } from '@/app/typesProduct'
 
 const Productss = () => {
 
-  const { products, fetchProducts, handleDelete } = useProductsForm()
+  const { products, productsDetails, fetchProducts, handleDelete } = useProductsForm()
   const { selectedIdProduct } = useAppContext()
 
   useEffect(() => {
     fetchProducts();
   }, [selectedIdProduct]);
 
+  if(!productsDetails) {
+    return <SpinnerComponent />
+  }
+
   return (
     <>
       <Box className='adminPagesBox categoriesBox '>
         <AdminHeader />
 
-        {products && products.map((product: Products) => (
-          <AdminCategoriesComponent
-            product={product}
-            key={product.id}
-            title={product.name}
-            handleDeleteProduct={() => handleDelete(product.id)}
-            handleDeleteCategory={() => { }}
-            imageUrl="/images/category/categories/defaultIcon.svg"
-          />
-        ))}
+        {products ? (
+          products.map((product: Products) => (
+            <AdminCategoriesComponent
+              product={product}
+              key={product.id}
+              title={product.name}
+              handleDeleteProduct={() => handleDelete(product.id)}
+              handleDeleteCategory={() => { }}
+              imageUrl="/images/category/categories/defaultIcon.svg"
+            />
+          ))
+        ) : (
+          <span>ЛАЛАЛАFKKKFFK</span>
+        )}
 
         <Link href='/admin/products/add'>
-          <Button className='buttonAdd button-position-categories' colorScheme='purple'>+</Button>
+          <Button
+            className='buttonAdd button-position-categories'
+            colorScheme='purple'
+          >
+            +
+          </Button>
         </Link>
       </Box>
     </>
